[appengine][xs]: revert app version in app.yaml to 1 (reason below).
We do this so that when we deploy app shows up as default/latest version -- it turns out that upgrading a deployed app to a new app version requires intervention via the appengine admin console. We should reserve version changes for genuinely significant upgrades (if we want to rollback because of errors can just use mercurial).