Simple fix for encoding problem.
Problem with encoding seems to have been related to the assumption that
it is possible to encode anything retrieved from the internet as UTF8
which isn't the case, particularly when the charset is something like
latin1 (128 from ascii plus some european chars up to 255).

After looking through the messytables code to find out how it handles
the encoding it appears that the UTF8Recoder there sets the charset
encoding to ignore utf8 chars that it can't convert (most of the latin1
charset >128, or so it appears).  As this is a preview, and after
discussion with DR, we decided that it is the most effective resolution
at present. Have copied the UTF8Recoder from messytables over the old
one that was part of the brewery code here.

So now when encountering non-utf8 charsets where it fails to convert
individual chars it will instead ignore those characters.  The option
does still exist to specify the encoding on the URL, so where something
like the spanish_example (from the messy tables test folder which is at
https://github.com/okfn/messytables/tree/master/horror) has chars dropped
it is possible to add encoding=latin1 to the end of the url and receive all
of the special characters.